CellIDTranslator: move most of the implementation to a source file.
