+++ include/deal.II/grid/cell_id_translator.h
@@ -14,13 +14,14 @@
 #define dealii_cell_id_translator_h
 
 #include <deal.II/base/exceptions.h>
-#include <deal.II/base/geometry_info.h>
+#include <deal.II/base/utilities.h>
 
 #include <deal.II/grid/cell_id.h>
 #include <deal.II/grid/tria_accessor.h>
 
 #include <cstdint>
 #include <limits>
+#include <vector>
 
 DEAL_II_NAMESPACE_OPEN
 
@@ -122,56 +123,6 @@ namespace internal
 
 
   template <int dim>
-  CellIDTranslator<dim>::CellIDTranslator(
-    const types::global_cell_index n_coarse_cells,
-    const types::global_cell_index n_global_levels)
-    : n_coarse_cells(n_coarse_cells)
-    , n_global_levels(n_global_levels)
-  {
-    // The class stores indices as types::global_cell_index variables,
-    // but when configuring deal.II with default flags, this is a 32-bit
-    // data type and it is possible with highly (locally) refined meshes
-    // that we exceed the maximal 32-bit numbers even with relatively
-    // modest numbers of cells. Check for this by first calculating
-    // the maximal index we will get in 64-bit arithmetic and testing
-    // that it is representable in 32-bit arithmetic:
-    std::uint64_t max_cell_index = 0;
-
-    for (unsigned int i = 0; i < n_global_levels; ++i)
-      max_cell_index +=
-        Utilities::pow<std::uint64_t>(ReferenceCells::max_n_children<dim>(),
-                                      i) *
-        n_coarse_cells;
-
-    max_cell_index -= 1;
-
-    AssertThrow(
-      max_cell_index <= std::numeric_limits<types::global_cell_index>::max(),
-      ExcMessage(
-        "You have exceeded the maximal number of possible indices this function "
-        "can handle. The current setup (n_coarse_cells=" +
-        std::to_string(n_coarse_cells) +
-        ", n_global_levels=" + std::to_string(n_global_levels) + ") requires " +
-        std::to_string(max_cell_index + 1) +
-        " indices but the current deal.II configuration only supports " +
-        std::to_string(std::numeric_limits<types::global_cell_index>::max()) +
-        " indices. You may want to consider to build deal.II with 64bit "
-        "indices (-D DEAL_II_WITH_64BIT_INDICES=\"ON\") to increase the limit "
-        "of indices."));
-
-    // Now do the whole computation again, but for real:
-    tree_sizes.reserve(n_global_levels + 1);
-    tree_sizes.push_back(0);
-    for (unsigned int i = 0; i < n_global_levels; ++i)
-      tree_sizes.push_back(tree_sizes.back() +
-                           Utilities::pow<types::global_cell_index>(
-                             ReferenceCells::max_n_children<dim>(), i) *
-                             n_coarse_cells);
-  }
-
-
-
-  template <int dim>
   types::global_cell_index
   CellIDTranslator<dim>::size() const
   {
@@ -225,52 +176,6 @@ namespace internal
     return (translate(cell) - tree_sizes[cell->level()]) *
              ReferenceCells::max_n_children<dim>() +
            i + tree_sizes[cell->level() + 1];
-  }
-
-
-
-  template <int dim>
-  CellId
-  CellIDTranslator<dim>::to_cell_id(const types::global_cell_index id) const
-  {
-    std::vector<std::uint8_t> child_indices;
-
-    types::global_cell_index id_temp = id;
-
-    types::global_cell_index level = 0;
-
-    for (; level < n_global_levels; ++level)
-      if (id < tree_sizes[level])
-        break;
-    level -= 1;
-
-    id_temp -= tree_sizes[level];
-
-    for (types::global_cell_index l = 0; l < level; ++l)
-      {
-        child_indices.push_back(id_temp %
-                                ReferenceCells::max_n_children<dim>());
-        id_temp /= ReferenceCells::max_n_children<dim>();
-      }
-
-    std::reverse(child_indices.begin(), child_indices.end());
-
-    return {id_temp, child_indices};
-  }
-
-
-
-  template <int dim>
-  types::global_cell_index
-  CellIDTranslator<dim>::to_level_cell_index(const CellId &cell_id)
-  {
-    // compute level id: c_{i+1} = c_{i}*(max_n_children) + q on path to cell
-    auto level_cell_id = cell_id.get_coarse_cell_id();
-    for (const auto &child_index : cell_id.get_child_indices())
-      level_cell_id =
-        level_cell_id * ReferenceCells::max_n_children<dim>() + child_index;
-
-    return level_cell_id;
   }
 } // namespace internal
