+++ source/grid/CMakeLists.txt
@@ -29,6 +29,7 @@ endif()
 
 set(_unity_include_src
   cell_id.cc
+  cell_id_translator.cc
   grid_refinement.cc
   intergrid_map.cc
   manifold.cc
@@ -69,6 +70,7 @@ setup_source_list("${_unity_include_src}"
 
 set(_inst
   cell_id.inst.in
+  cell_id_translator.inst.in
   grid_generator_cgal.inst.in
   grid_generator_from_name.inst.in
   grid_generator.inst.in
